wifi: Fix connman scan wifi failure after start
Depending on how the supplicant is configured there is a case where
connman is unable to scan right after starting.

The sequence that leads to that failure is as simple as:

	$ connmanctl enable wifi
	Enabled wifi
	$ connmanctrl scan wifi
	Error /net/connman/technology/wifi: No carrier

The problem is located on the enable wifi command, where the
creation callback is never called.

After the patch scan works fine.

	$ connmanctrl scan wifi
	Scan completed for wifi

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>